SineTables.jl Documentation

SineTables.SineTable — Type

Store the lookup table and additional parameters

SineTables.create_sine_table — Method

create_sine_table(::Type{T}, points::Integer, bias::Number=0, scale::Number=1) -> SineTable{T}

Creates a lookup table for the sine function for the entire wave.

SineTables.get_index — Method

get_index(phase::Number, points::Integer) -> Integer

Utility function to fetch the index corresponding to the phase

SineTables.lut_cosine! — Method

lut_cosine!(result::T, table::SineTable, phase::Float64) -> nothing

Mutating version of lut_cosine, stores the cosine value in result.

SineTables.lut_cosine — Method

lut_cosine(table::SineTable, phase::Float64) -> T

Fetch a cosine value for a given phase using the lookup table.

SineTables.lut_sine! — Method

lut_sine!(result::T, table::SineTable, phase::Float64) -> nothing

Mutating version of lut_sine, stores the sine value in result.

SineTables.lut_sine — Method

lut_sine(table::SineTable, phase::Float64) -> T

Fetch a sine value for a given phase using the lookup table.